+++ lib/ts/utils.ts
@@ -4,6 +4,7 @@ import NormalisedURLDomain from "./normalisedURLDomain";
 import NormalisedURLPath from "./normalisedURLPath";
 import type { BaseRequest, BaseResponse } from "./framework";
 import { logDebugMessage } from "./logger";
+import { APIResponseGeneralError } from "./types";
 
 export function getLargestVersionFromIntersection(v1: string[], v2: string[]): string | undefined {
     let intersection = v1.filter((value) => v2.indexOf(value) !== -1);
@@ -119,4 +120,11 @@ export function humaniseMilliseconds(ms: number): string {
         if (h > 1) suffix = "s";
         return `${h} hour${suffix}`;
     }
+}
+
+export function convertToAPIResponseGeneralError(err: any): APIResponseGeneralError {
+    return {
+        status: "GENERAL_ERROR",
+        message: err.message !== undefined ? err.message : "Something went wrong",
+    };
 }